Add code to check parameters
If the input mount point cannot be found, the program will print a prompt message and exit normally without causing an exception that may affect other processes in the program.

If the program unable to connect to a node in the cluster(according to the input), or the connection has been established, but the shared storage disk is not found, the prompt message will also be printed and the program will be terminated normally.