delete GEN, fix decode entry , add -d
Delete the GEN field that was originally displayed

Since the kernel has two output modes for the 'name' field, it needs to be parsed in different ways. When parsing the name of the dentry, the last 8 bytes store the inode node number. The other entry is the 16-byte save inode number from 7 to 23. Fixed this parsing bug

use id in the command to show the system inode number infomation