Add dynamic display method
According to the number of rows and columns of the terminal window, the number of displayed lines is dynamically adjusted, and if the terminal window size changes during the program running process, o2locktop dynamically adjusts the displayed number of lines. And the -l parameters is still available.

In ‘detial’ mode, no extra information is displayed, but when saving log files, all information is saved.

Hide the cursor when displayed