Use the producer–consumer pattern to accelerate o2locktop
Please be note that because python GIL, this change can only accelerate the process when the network is slow.
The pattern is like that when the 1st time get the data from cluster, a thread wait for the data and process it. but when the 1st the process the data, and we can start to get the next time data from server.
When the condition of network is pretty good, this update can not accelerate the process. Because read data from network and process the data are both CPU intensive.